Create README.md
Our research is devoted to systematic stock risk in Russia. The research is based on the CAPM regression model and its application to the Russian market. The research consists of 4 relatively independent parts:

1. The best indicator for CAPM
2. Structural shifts in the stock market
3. Dependence of changes in the MOEX index
4. Influence of commodity and currency indices

As a result, we formulated several conclusions that, in our opinion, can be useful for investment analysts and financiers.